+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Dodo x.x.x (not yet released)
-* Fix doctype creation in DOMParser.
+* Fix doctype creation and "significant whitespace" handling in DOMParser.
 
 # Dodo 0.1.0 (2021-07-04)
 Initial release.